add compare to navbar
