+++ components/DataCompareListing.tsx
@@ -1,45 +1,19 @@
 import { useEffect, useState } from "react";
-import { useSearchParams, useRouter } from "next/navigation";
 import styles from "@/styles/DataCompare.module.css"; // New CSS module for the page
 import InformationShort from "@/components/InformationShort";
 import TrajectoryVisualizerDual from "@/components/TrajectoryVisualizerDual";
 import { SharedState } from "@/types/pageInterface";
 import { useInfoData } from "@/utils/useInfoData";
 
-const DataCompareListing: React.FC = () => {
-  const searchParams = useSearchParams();
-  const router = useRouter();
+interface DataCompareListingProps {
+  data1: string;
+  data2: string;
+}
 
-  // Get the query parameters and set defaults
-  const initialData1 = searchParams.get("data1") || "default data 1";
-  const initialData2 = searchParams.get("data2") || "default data 2";
-
-  // Set up state for data1 and data2
-  const [data1, setData1] = useState(initialData1);
-  const [data2, setData2] = useState(initialData2);
-
-  // Function to update the URL search params
-  const updateSearchParams = (key: string, value: string) => {
-    const currentParams = new URLSearchParams(searchParams.toString());
-    currentParams.set(key, value);
-
-    router.replace(`?${currentParams.toString()}`);
-  };
-
-  // Handle input change for data1
-  const handleData1Change = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const newValue = e.target.value;
-    setData1(newValue);
-    updateSearchParams("data1", newValue);
-  };
-
-  // Handle input change for data2
-  const handleData2Change = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const newValue = e.target.value;
-    setData2(newValue);
-    updateSearchParams("data2", newValue);
-  };
-
+const DataCompareListing: React.FC<DataCompareListingProps> = ({
+  data1,
+  data2,
+}) => {
   const infoData1 = useInfoData(data1);
   const infoData2 = useInfoData(data2);
 
@@ -98,36 +72,12 @@ const DataCompareListing: React.FC = () => {
         <div className={styles.column}>
           <div className={styles.container}>
             {/* First Column */}
-            <div className={styles.inputGroup}>
-              <label htmlFor="data1" className={styles.label}>
-                Data 1:{" "}
-              </label>
-              <input
-                id="data1"
-                type="text"
-                value={data1}
-                onChange={handleData1Change}
-                className={styles.input}
-              />
-            </div>
             <InformationShort folderName={data1} />
           </div>
         </div>
         <div className={styles.column}>
           <div className={styles.container}>
             {/* Second Column */}
-            <div className={styles.inputGroup}>
-              <label htmlFor="data2" className={styles.label}>
-                Data 2:{" "}
-              </label>
-              <input
-                id="data2"
-                type="text"
-                value={data2}
-                onChange={handleData2Change}
-                className={styles.input}
-              />
-            </div>
             <InformationShort folderName={data2} />
           </div>
         </div>
